Move the CheckerBoard class to verde.synthetic
