+++ verde/tests/test_scipy.py
@@ -15,7 +15,7 @@ import pandas as pd
 import pytest
 
 from ..coordinates import grid_coordinates
-from ..datasets.synthetic import CheckerBoard
+from ..synthetic import CheckerBoard
 from ..scipygridder import ScipyGridder
